+++ src/printer.rs
@@ -12,7 +12,9 @@ use std::io::{self, Write};
 pub enum ReportFormat {
     Json,
     JsonPretty,
+    // TODO: add CSV
     Markdown,
+    UserMarkdown,
     Default,
 }
 
@@ -29,13 +31,14 @@ impl TodoR {
         Ok(())
     }
 
-    /// Writes TODOs in TodoR serialized in a markdown format
+    /// Writes TODOs in TodoR serialized in a markdown table format.
+    /// Tables are organized by TODO tag type.
     fn write_markdown(&self, out_buffer: &mut impl Write) -> Result<(), Error> {
-        let mut tag_tables: FnvHashMap<String, String> = FnvHashMap::default();
+        let mut tag_tables: FnvHashMap<&str, String> = FnvHashMap::default();
 
         for ptodo in self.iter() {
             let todo = ptodo.todo;
-            let table_string = tag_tables.entry(todo.tag.clone()).or_insert_with(|| {
+            let table_string = tag_tables.entry(&todo.tag).or_insert_with(|| {
                 format!(
                     "### {}s\n| Filename | line | {} |\n|:---|:---:|:---|\n",
                     todo.tag, todo.tag,
@@ -53,6 +56,61 @@ impl TodoR {
 
         for table_strings in tag_tables.values() {
             writeln!(out_buffer, "{}", table_strings)?;
+        }
+
+        Ok(())
+    }
+
+    /// Writes TODOs in TodoR serialized in a markdown table format.
+    /// Tables are organized by TODO user.
+    fn write_user_markdown(&self, out_buffer: &mut impl Write) -> Result<(), Error> {
+        let mut user_tables: FnvHashMap<&str, String> = FnvHashMap::default();
+        let mut untagged_todos_string = "".to_string();
+
+        for ptodo in self.iter() {
+            let todo = ptodo.todo;
+            let users = todo.users();
+
+            if users.is_empty() {
+                writeln!(
+                    untagged_todos_string,
+                    "| {} | {} | {} | {} |",
+                    ptodo.file.display(),
+                    todo.line,
+                    todo.tag,
+                    todo.content,
+                )?;
+            } else {
+                for user in users {
+                    let table_string = user_tables.entry(user).or_insert_with(|| {
+                        format!(
+                            "### {}\n| Filename | line | type | content |\n|:---|:---:|:---:|:---|\n",
+                            user,
+                        )
+                    });
+
+                    writeln!(
+                        table_string,
+                        "| {} | {} | {} | {} |",
+                        ptodo.file.display(),
+                        todo.line,
+                        todo.tag,
+                        todo.content,
+                    )?;
+                }
+            }
+        }
+
+        for table_strings in user_tables.values() {
+            writeln!(out_buffer, "{}", table_strings)?;
+        }
+
+        if !untagged_todos_string.is_empty() {
+            writeln!(
+                out_buffer,
+                "### Untagged\n| Filename | line | type | content |\n|:---|:---:|:---:|:---|\n{}",
+                untagged_todos_string,
+            )?;
         }
 
         Ok(())
@@ -78,6 +136,8 @@ impl TodoR {
             ReportFormat::Json => TodoR::write_json,
             ReportFormat::JsonPretty => TodoR::write_pretty_json,
             ReportFormat::Markdown => TodoR::write_markdown,
+            ReportFormat::UserMarkdown => TodoR::write_user_markdown,
+            // TODO: make default print have no colors
             ReportFormat::Default => TodoR::write_todos,
         };